fail info| Puts the new check_output code-path under test
To achieve this we redirect stdout so we can capture it and compare it
against expected. We also mock out the check_output function so that it
always raises the expected exception type.

We also set and reset the check_output mock during the test so it
doesn't impact other tests